Struct cdc::CdcObserver

pub struct CdcObserver {
    sched: Scheduler<Task>,
    observe_regions: Arc<RwLock<HashMap<u64, ObserveID>>>,
    cmd_batches: RefCell<Vec<CmdBatch>>,
}

An Observer for CDC.

It observes raftstore internal events, such as:

Raft role change events,

Apply command events.

Fields
sched: Scheduler<Task>observe_regions: Arc<RwLock<HashMap<u64, ObserveID>>>cmd_batches: RefCell<Vec<CmdBatch>>
Implementations

impl CdcObserver

pub fn new(sched: Scheduler<Task>) -> CdcObserver

Create a new CdcObserver.

Events are strong ordered, so sched must be implemented as a FIFO queue.

pub fn register_to(&self, coprocessor_host: &mut CoprocessorHost<RocksEngine>)

pub fn subscribe_region(
&self,
region_id: u64,
observe_id: ObserveID
) -> Option<ObserveID>

Subscribe an region, the observer will sink events of the region into its scheduler.

Return pervious ObserveID if there is one.

pub fn unsubscribe_region(
&self,
region_id: u64,
observe_id: ObserveID
) -> Option<ObserveID>

Stops observe the region.

Return ObserverID if unsubscribe successfully.

pub fn is_subscribed(&self, region_id: u64) -> Option<ObserveID>

Check whether the region is subscribed or not.